requirements.txt
appnope==0.1.3
asgiref==3.8.1
asttokens==2.2.1
autopep8==2.0.4
backcall==0.2.0
certifi==2024.12.14
charset-normalizer==3.4.2
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.9
executing==1.2.0
Faker==37.3.0
filelock==3.16.1
hubspot-api-client==12.0.0
idna==3.10
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
Pygments==2.15.1
python-dateutil==2.8.2
pyzmq==25.1.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.28.0
wcwidth==0.2.6
